// back-end/user/src/lib/@types/user.type.ts
export type RegisterResponse = {
  status: number;
  failed: boolean;
  message: string;
};

export type userUpdate = {
  name: string;
  email: string;
  user_discord: string;
  user_reference: string;
  phone: string;
};


export type userUpdatePassword = {
  email: string;
  user_reference: string;
  password: string;
  confirm_password: string;
};